# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ResuMate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WorkExperience {
  id: string;
  company: string;
  role: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  description: string;
  current: boolean;
}

export interface Education {
  id: string;
  institution: string;
  degree: string;
  field: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  current: boolean;
}

export type ResumeDataTemplate = typeof ResumeDataTemplate[keyof typeof ResumeDataTemplate];


export const ResumeDataTemplate = {
  classic: 'classic',
  modern: 'modern',
} as const;

export interface ResumeData {
  fullName: string;
  email: string;
  phone: string;
  location: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  linkedin?: string | null;
  summary: string;
  workExperience: WorkExperience[];
  education: Education[];
  skills: string[];
  template: ResumeDataTemplate;
  isPremium: boolean;
}

export interface Resume {
  id: number;
  email: string;
  data: ResumeData;
  updatedAt: string;
}

export interface ResumeResponse {
  resume: Resume | null;
}

export interface ResumeInput {
  email: string;
  data: ResumeData;
}

export type LookupResumeParams = {
email: string;
};