# lib/api-zod/src/generated/types/education.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ResuMate API
 * OpenAPI spec version: 0.1.0
 */

export interface Education {
  id: string;
  institution: string;
  degree: string;
  field: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  current: boolean;
}