# lib/api-zod/src/generated/types/resume.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ResuMate API
 * OpenAPI spec version: 0.1.0
 */
import type { ResumeData } from './resumeData';

export interface Resume {
  id: number;
  email: string;
  data: ResumeData;
  updatedAt: string;
}