# lib/api-zod/src/generated/types/resumeData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ResuMate API
 * OpenAPI spec version: 0.1.0
 */
import type { Education } from './education';
import type { ResumeDataTemplate } from './resumeDataTemplate';
import type { WorkExperience } from './workExperience';

export interface ResumeData {
  fullName: string;
  email: string;
  phone: string;
  location: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  linkedin?: string | null;
  summary: string;
  workExperience: WorkExperience[];
  education: Education[];
  skills: string[];
  template: ResumeDataTemplate;
  isPremium: boolean;
}