# lib/api-zod/src/generated/types/resumeResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ResuMate API
 * OpenAPI spec version: 0.1.0
 */
import type { Resume } from './resume';

export interface ResumeResponse {
  resume: Resume | null;
}