# lib/api-zod/src/generated/types/workExperience.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ResuMate API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkExperience {
  id: string;
  company: string;
  role: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  description: string;
  current: boolean;
}